Further rename timeseries_api & adapter -> adapters
