+++ src/main/scala/org/wdias/adapters/metadata_adapter/models/Parameters.scala
@@ -1,4 +1,4 @@
-package org.wdias.adapter.metadata_adapter.models
+package org.wdias.adapters.metadata_adapter.models
 
 import org.wdias.constant.ParameterType.ParameterType
 import slick.ast.BaseTypedType
