+++ src/main/scala/org/wdias/adapters/metadata_adapter/models/TimeSteps.scala
@@ -1,4 +1,4 @@
-package org.wdias.adapter.metadata_adapter.models
+package org.wdias.adapters.metadata_adapter.models
 
 import org.wdias.constant.TimeStepUnit
 import org.wdias.constant.TimeStepUnit.TimeStepUnit
