+++ src/main/scala/org/wdias/adapters/vector_adapter/VectorAdapter.scala
@@ -1,4 +1,4 @@
-package org.wdias.adapter.grid_adapter
+package org.wdias.adapters.vector_adapter
 
 import java.time.format.DateTimeFormatter
 import java.time.{LocalDateTime, ZoneId}
@@ -7,7 +7,7 @@ import akka.actor.{Actor, ActorIdentity, ActorLogging, ActorRef, Identify}
 import akka.pattern.pipe
 import akka.util.Timeout
 import com.paulgoldbaum.influxdbclient.Parameter.Precision
-import org.wdias.adapter.grid_adapter.GridAdapter._
+import org.wdias.adapters.vector_adapter.VectorAdapter._
 import org.wdias.extensions.ExtensionHandler.ExtensionHandlerData
 import ucar.ma2.DataType
 import ucar.nc2.{Attribute, Dimension}
@@ -21,7 +21,7 @@ import org.wdias.constant._
 import scala.concurrent.ExecutionContext.Implicits.global
 import scala.concurrent.duration._
 
-object GridAdapter {
+object VectorAdapter {
 
   case class StoreTimeSeries(timeSeriesEnvelop: TimeSeriesEnvelop)
   case class StoreValidatedTimeSeries(timeSeriesEnvelop: TimeSeriesEnvelop)
@@ -36,8 +36,7 @@ object GridAdapter {
 
 }
 
-class GridAdapter extends Actor with ActorLogging {
-
+class ScalarAdapter extends Actor with ActorLogging {
 
   implicit val timeout: Timeout = Timeout(15 seconds)
 
@@ -52,7 +51,7 @@ class GridAdapter extends Actor with ActorLogging {
     val records: List[Record] = result.series.head.records
     records.foreach { record =>
       log.info(record.allValues.toString())
-      val dateTimeStr: String = record.allValues.head.toString.split('Z')(0)
+      val dateTimeStr: String = record.allValues(0).toString.split('Z')(0)
       val dateTime = LocalDateTime.parse(dateTimeStr)
       val value: Double = record.allValues(valueIndex).toString.toDouble
       points = points :+ DataPoint(dateTime.format(formatter), value)
@@ -66,7 +65,7 @@ class GridAdapter extends Actor with ActorLogging {
     val location = "/tmp/testWrite.nc"
 
     if(Files.exists(Paths.get(location))) {
-      return
+      return null
     }
     println("File does not exists. Create new ", location)
     import ucar.nc2.NetcdfFileWriter
@@ -113,6 +112,7 @@ class GridAdapter extends Actor with ActorLogging {
       case e: Exception => System.err.printf("ERROR creating file %s%n%s", location, e.getMessage);
     }
     writer.close()
+    null
   }
 
   def receive: Receive = {
