+++ src/main/scala/org/wdias/export/csv/ExportCSV.scala
@@ -2,7 +2,7 @@ package org.wdias.export.csv
 
 import akka.actor.{Actor, ActorIdentity, ActorLogging, ActorRef, Identify}
 import akka.util.Timeout
-import org.wdias.adapter.scalar_adapter.ScalarAdapter.GetTimeSeries
+import org.wdias.adapters.scalar_adapter.ScalarAdapter.GetTimeSeries
 import org.wdias.constant.MetaData
 
 import scala.concurrent.duration._
@@ -16,7 +16,7 @@ class ExportCSV extends Actor with ActorLogging {
   implicit val timeout: Timeout = Timeout(15 seconds)
 
   var adapterRef: ActorRef = _
-  context.actorSelection("/user/adapter") ! Identify(None)
+  context.actorSelection("/user/scalarAdapter") ! Identify(None)
 
   def receive = {
     case ExportCSVFile(metaData) =>
