+++ src/main/scala/org/wdias/import/csv/ImportCSV.scala
@@ -5,7 +5,7 @@ import java.time.{LocalDateTime, ZoneId}
 
 import akka.actor.{Actor, ActorIdentity, ActorLogging, ActorRef, Identify}
 import akka.util.Timeout
-import org.wdias.adapter.scalar_adapter.ScalarAdapter.StoreTimeSeries
+import org.wdias.adapters.scalar_adapter.ScalarAdapter.StoreTimeSeries
 import org.wdias.constant._
 
 import scala.concurrent.duration._
@@ -23,7 +23,7 @@ class ImportCSV extends Actor with ActorLogging {
   implicit val timeout: Timeout = Timeout(15 seconds)
 
   var adapterRef: ActorRef = _
-  context.actorSelection("/user/adapter") ! Identify(None)
+  context.actorSelection("/user/scalarAdapter") ! Identify(None)
 
   def receive: Receive = {
     case ImportCSVFile(metaData, source) =>
