+++ src/main/scala/org/wdias/timeseries_api/TimeseriesAPIRoutes.scala
@@ -10,7 +10,7 @@ import akka.http.scaladsl.server.directives.RouteDirectives.complete
 import akka.pattern.ask
 import akka.stream.ActorMaterializer
 import akka.util.Timeout
-import org.wdias.adapter.scalar_adapter.ScalarAdapter.Result
+import org.wdias.adapters.scalar_adapter.ScalarAdapter.Result
 import org.wdias.constant.{MetaData, Protocols}
 import org.wdias.export.json.ExportJSON.ExportJSONData
 import org.wdias.input.InputRoutes
@@ -19,7 +19,7 @@ import scala.concurrent.Future
 import scala.concurrent.duration._
 
 // REST API Routes class
-trait RESTAPIRoutes extends Protocols {
+trait TimeseriesAPIRoutes extends Protocols {
   // abstract system value will be provide by app
   implicit def system: ActorSystem
   implicit def materializer: ActorMaterializer
